update cmd package w/ makefile improvments
