--- {{cookiecutter.name}}/main.go
@@ -1,7 +1,0 @@
-package main
-
-import "{{cookiecutter.name}}/cmd"
-
-func main() {
-	cmd.New().Execute()
-}
